requirements.txt
dj-database-url==0.5.0
Django==3.0.4
django-heroku==0.3.1
gunicorn==20.0.4
mysqlclient==1.4.6
psycopg2==2.8.4
python-dotenv==0.12.0
requests==2.22.0
sqlparse==0.3.1
unicodecsv==0.14.1
urllib3==1.25.8
whitenoise==5.0.1
celery==4.4.2
django_celery_results==1.2.1
django-crispy-forms==1.9.0